Fix problem with negative port numbers. There is a mailing list thread describing the problem here: http://sourceforge.net/mailarchive/forum.php?thread_name=46EBD7B4.8070506%40gmail.com&forum_name=iaxclient-devel
git-svn-id: svn://svn.code.sf.net/p/iaxclient/code/trunk@1177 f854266a-e20d-0410-a9be-8852184c940f

+++ lib/iaxclient.h
@@ -22,7 +22,7 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-	
+
 /*!
   \file iaxclient.h
   \brief The IAXClient API
@@ -635,7 +635,7 @@ EXPORT void iaxc_set_preferred_source_udp_port(int sourceUdpPort);
 
 	\return The UDP port bound to; -1 if no port or
 */
-EXPORT short iaxc_get_bind_port();
+EXPORT int iaxc_get_bind_port();
 
 /*!
 	Initializes the IAXClient library
